gpu_unai,gpu_lib: Fix VRAM framebuffer alignment causing screen combing
Alignment of start of framebuffer VRAM needs to be 16 bytes
(8 pixels) to work with gpu_unai's blit_mask pixel-skipping.
Fixes intermittent vertical comb effect.

Aligned gpulib vram to 8192 bytes on GCW Zero platform to minimize
TLB fills required to traverse it.

NOTE: this has nothing to do with normal artifacts resulting from
using gpu_unai's pixel skipping, like pause screen of Crash Team
Racing.

+++ src/gpu/gpu_unai/gpu_unai.h
@@ -220,10 +220,10 @@ struct gpu_unai_t {
 	//u32   tInc, tMsk;
 	u32   u4_msk, v4_msk;
 
-	u32 blit_mask;          // Determines what pixels to skip when rendering.
+	u8 blit_mask;           // Determines what pixels to skip when rendering.
 	                        //  Only useful on low-resolution devices using
 	                        //  a simple pixel-dropping downscaler for PS1
-                            //  high-res modes. See 'pixel_skip' option.
+	                        //  high-res modes. See 'pixel_skip' option.
 
 	u8 ilace_mask;          // Determines what lines to skip when rendering.
 	                        //  Normally 0 when PS1 240 vertical res is in
